Community monitor: separate modal manager from modal
