+++ front/app/containers/App/ModalQueue/modals/CommunityMonitor/ModalManager.test.tsx
